requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3
bcrypt==4.0.1 